Rename the merged dock to Inspector and consolidate the tree code
The merged panel presents as Inspector, matching what it does rather
than the old entity-tree name, and the menu-bar test asserts the
distinct View/Panels entries in their declared order. The superseded
MeshInfo and EntityTreePanel feature classes are removed together with
the structure-only and duplicated tests, while the ported test cases
stay verbatim from their originals. The world poll timer moves into the
entity tree widget, the only tree that needs polling, running just while
that tree is on screen. The tree base, both tree widgets, and the
unified dock consolidate into _tree_panel.py, where a single module ends
the import cycle that once forced a separate leaf module for the base.
